refactor(api): move update referer usecase to team context
