+++ api/tests/team/unit/domain/usecases/update-certification-center-referer_test.js
@@ -1,5 +1,5 @@
-import { updateCertificationCenterReferer } from '../../../../lib/domain/usecases/update-certification-center-referer.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { updateCertificationCenterReferer } from '../../../../../src/team/domain/usecases/update-certification-center-referer.js';
+import { domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
 describe('Unit | UseCase | update-certification-center-referer', function () {
   context('when there is already a referer', function () {
